feat(upscale): arbitrary scale factors, so a plan can ask for 1.3x
The 2/4/6 ladder was my assumption, not a Topaz limit. The Autopilot registry
factor takes arbitrary decimals — verified twice: 1.3 on a 3660x2997 source
produced exactly 4758x3896, and 1.3 on a 6528x4352 photo produced 8486x5658.

This matters because print plans rarely need a whole factor. Sweeping every
blocked arrangement of one 28 MP photo, the required factors ran 1.1x to 1.9x
and NOT ONE needed 2x. Rounding those up meant 2.4x the pixels — every one
carrying synthesized texture the plan never asked for, and 2.4x the file. The
real 1.3x export above is 13.6 MB where 2x was 33 MB.

scale is now a float in [1.05, 6.0]; SCALES is only the list of UI presets. The
/split diagnostic asks for exactly what the plan needs and says so ("That is
the exact factor this plan needs — 8,160 x 5,440 px"), with upscaleFor rounding
up to a tenth so the result clears the floor rather than landing under it.

Whole factors still render as integers in filenames (-2x, not -2.0x) and
fractional ones keep their decimal (-1.3x), round-tripping back through
_parse_variant so the prior-exports list and the Split source picker read them
correctly.

+++ photosearch/admin_api.py
@@ -946,7 +946,9 @@ def admin_rerun_passes(req: RerunRequest):
 class UpscaleRequest(BaseModel):
     photo_ids: list[int]
     # None = let Topaz Autopilot choose (and never touch the user's settings).
-    scale: Optional[int] = None
+    # Any decimal in [1.05, 6]: the planner asks for exactly the factor a
+    # plan needs rather than rounding up to a whole number.
+    scale: Optional[float] = None
     enhancements: list[str] = ["upscale"]
     # None = Autopilot's own pick (usually "High Fidelity V2").
     model: Optional[str] = None
